Extract generic ScopeRenderer trait into top-level renderer module
Move the renderer trait (renamed from ClassedRenderer to ScopeRenderer)
and its driving loop into src/renderer.rs with no HTML dependencies.
HTML-specific renderers (HtmlClassedRenderer, LineHighlightingRenderer)
live in src/html/renderer.rs as specializations. DefaultClassedRenderer
is now a deprecated alias for HtmlClassedRenderer.
